Remove top-level `__init__.py` files from library and driver
This leads to cleaner module imports.
Also provide the `skip_without_gripper` decorator in the `utility`
module so that the driver can use it.
