+++ schunk_gripper_library/tests/conftest.py
@@ -1,10 +1,8 @@
 import pytest
-from schunk_gripper_library.tests.etc.pseudo_terminals import Connection
-from schunk_gripper_library.tests.etc.modbus_server import ModbusServer
+from .etc.pseudo_terminals import Connection
+from .etc.modbus_server import ModbusServer
 import asyncio
 import threading
-from pathlib import Path
-from httpx import Client, ConnectTimeout, ConnectError
 
 
 @pytest.fixture(scope="module")
@@ -48,22 +46,3 @@ def modbus_server(pseudo_terminals):
     loop.call_soon_threadsafe(loop.stop)
     thread.join()
     loop.close()
-
-
-def gripper_available():
-    client = Client()
-    try:
-        webserver_up = client.get(
-            "http://0.0.0.0:8000/adi/data.json", timeout=1.0
-        ).is_success
-    except (ConnectTimeout, ConnectError):
-        webserver_up = False
-    modbus_server_up = Path("/dev/ttyUSB0").exists() or ""
-    if webserver_up and modbus_server_up:
-        return True
-    return False
-
-
-skip_without_gripper = pytest.mark.skipif(
-    not gripper_available(), reason="No gripper/simulator available"
-)
