+++ schunk_gripper_library/tests/test_control_operations.py
@@ -1,4 +1,4 @@
-from ..schunk_gripper_library.driver import Driver
+from schunk_gripper_library.driver import Driver
 import pytest
